chat-api/backend/requirements.txt
# Core AWS dependencies
boto3>=1.34.0
botocore>=1.34.0

# Authentication and JWT
PyJWT>=2.8.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# HTTP requests
requests>=2.32.5

# Utilities
python-dateutil>=2.8.0